Change folder structure
Starting clean code architecture
