+++ code-challenge-android-kotlin/app/src/main/AndroidManifest.xml
@@ -21,7 +21,7 @@
             </intent-filter>
         </activity>
 
-        <activity android:name=".home.HomeActivity" />
+        <activity android:name=".home.presentation.ui.HomeActivity" />
 
     </application>
