+++ code-challenge-android-kotlin/app/src/main/java/com/arctouch/codechallenge/base/presentation/ui/BaseActivity.kt
@@ -1,7 +1,7 @@
-package com.arctouch.codechallenge.base
+package com.arctouch.codechallenge.base.presentation.ui
 
 import androidx.appcompat.app.AppCompatActivity
-import com.arctouch.codechallenge.api.TmdbApi
+import com.arctouch.codechallenge.home.presentation.network.TmdbApi
 import okhttp3.OkHttpClient
 import retrofit2.Retrofit
 import retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory
