--- code-challenge-android-kotlin/app/src/main/java/com/arctouch/codechallenge/data/Cache.kt
@@ -1,12 +1,0 @@
-package com.arctouch.codechallenge.data
-
-import com.arctouch.codechallenge.model.Genre
-
-object Cache {
-
-    var genres = listOf<Genre>()
-
-    fun cacheGenres(genres: List<Genre>) {
-        this.genres = genres
-    }
-}
