+++ code-challenge-android-kotlin/app/src/main/java/com/arctouch/codechallenge/home/data/datasources/Cache.kt
@@ -1,0 +1,12 @@
+package com.arctouch.codechallenge.home.data.datasources
+
+import com.arctouch.codechallenge.home.domain.entities.Genre
+
+object Cache {
+
+    var genres = listOf<Genre>()
+
+    fun cacheGenres(genres: List<Genre>) {
+        Cache.genres = genres
+    }
+}
