+++ code-challenge-android-kotlin/app/src/main/java/com/arctouch/codechallenge/home/domain/entities/Genre.kt
@@ -1,0 +1,3 @@
+package com.arctouch.codechallenge.home.domain.entities
+
+data class Genre(val id: Int, val name: String)
