+++ code-challenge-android-kotlin/app/src/main/java/com/arctouch/codechallenge/home/domain/entities/Movie.kt
@@ -1,0 +1,15 @@
+package com.arctouch.codechallenge.home.domain.entities
+
+import com.squareup.moshi.Json
+
+
+data class Movie(
+        val id: Int,
+        val title: String,
+        val overview: String?,
+        val genres: List<Genre>?,
+        @Json(name = "genre_ids") val genreIds: List<Int>?,
+        @Json(name = "poster_path") val posterPath: String?,
+        @Json(name = "backdrop_path") val backdropPath: String?,
+        @Json(name = "release_date") val releaseDate: String?
+)
