+++ code-challenge-android-kotlin/app/src/main/java/com/arctouch/codechallenge/home/presentation/network/TmdbApi.kt
@@ -1,8 +1,8 @@
-package com.arctouch.codechallenge.api
+package com.arctouch.codechallenge.home.presentation.network
 
-import com.arctouch.codechallenge.model.GenreResponse
-import com.arctouch.codechallenge.model.Movie
-import com.arctouch.codechallenge.model.UpcomingMoviesResponse
+import com.arctouch.codechallenge.home.domain.entities.Movie
+import com.arctouch.codechallenge.home.presentation.network.models.GenreResponse
+import com.arctouch.codechallenge.home.presentation.network.models.UpcomingMoviesResponse
 import io.reactivex.Observable
 import retrofit2.http.GET
 import retrofit2.http.Path
