+++ code-challenge-android-kotlin/app/src/main/java/com/arctouch/codechallenge/home/presentation/network/models/GenreResponse.kt
@@ -1,0 +1,7 @@
+package com.arctouch.codechallenge.home.presentation.network.models
+
+import com.arctouch.codechallenge.home.domain.entities.Genre
+
+
+data class GenreResponse(val genres: List<Genre>)
+
