+++ code-challenge-android-kotlin/app/src/main/java/com/arctouch/codechallenge/home/presentation/ui/HomeActivity.kt
@@ -1,11 +1,12 @@
-package com.arctouch.codechallenge.home
+package com.arctouch.codechallenge.home.presentation.ui
 
 import android.os.Bundle
 import android.view.View
 import com.arctouch.codechallenge.R
-import com.arctouch.codechallenge.api.TmdbApi
-import com.arctouch.codechallenge.base.BaseActivity
-import com.arctouch.codechallenge.data.Cache
+import com.arctouch.codechallenge.home.presentation.network.TmdbApi
+import com.arctouch.codechallenge.base.presentation.ui.BaseActivity
+import com.arctouch.codechallenge.home.data.datasources.Cache
+import com.arctouch.codechallenge.home.presentation.ui.adapter.HomeAdapter
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.schedulers.Schedulers
 import kotlinx.android.synthetic.main.home_activity.*
