+++ code-challenge-android-kotlin/app/src/main/java/com/arctouch/codechallenge/splash/SplashActivity.kt
@@ -3,10 +3,10 @@ package com.arctouch.codechallenge.splash
 import android.content.Intent
 import android.os.Bundle
 import com.arctouch.codechallenge.R
-import com.arctouch.codechallenge.api.TmdbApi
-import com.arctouch.codechallenge.base.BaseActivity
-import com.arctouch.codechallenge.data.Cache
-import com.arctouch.codechallenge.home.HomeActivity
+import com.arctouch.codechallenge.home.presentation.network.TmdbApi
+import com.arctouch.codechallenge.base.presentation.ui.BaseActivity
+import com.arctouch.codechallenge.home.data.datasources.Cache
+import com.arctouch.codechallenge.home.presentation.ui.HomeActivity
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.schedulers.Schedulers
