+++ code-challenge-android-kotlin/app/src/main/java/com/arctouch/codechallenge/util/MovieImageUrlBuilder.kt
@@ -1,6 +1,6 @@
 package com.arctouch.codechallenge.util
 
-import com.arctouch.codechallenge.api.TmdbApi
+import com.arctouch.codechallenge.home.presentation.network.TmdbApi
 
 private val POSTER_URL = "https://image.tmdb.org/t/p/w154"
 private val BACKDROP_URL = "https://image.tmdb.org/t/p/w780"
